requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.1.0
django-cors-headers==4.2.0